Teljes kódbázis-audit 1. hullám: fordítás-blokkoló + exploit + Folia + dead code
KRITIKUS:
- 4 spell (Whirlwind/DeepBreath/SpinningCraneKick/FlyingSerpentKick): statikus
  helper instance getId()-t hívott — nem fordult; de-static
- RelicManager: a relog nem nullázza többé a relikvia-képesség cooldownját
- WorldBossManager: jutalom csak a követett bossért (crash-árva nem fizet duplán)
- RitualManager: üres sacrifice-lista = nincs áldozat-követelmény — a pakt_oltar
  (E25) eddig egyáltalán nem volt elvégezhető
- EscortManager: EventSpawnGuard bekötve (konvoj + hullám; új escort mátrix-sor)

KÖZEPES:
- AchievementManager: vagyon = összes valuta összege (default-valutás olvasás a
  nem-NEUTRAL játékosokat kizárta)
- KingManager: csak aktuális frakciótag koronázható (átállt szavazó nem)
- ClaimManager: oszlop-számítás long (int-túlcsordulásos plafon-kerülés zárva)
- HonorDuel accept: kihívó-PDC írás a kihívó saját szálán (Folia) + kimenő
  felkérés expiry-párjának takarítása kilépéskor
- GuildManager.members: CopyOnWriteArrayList (CME több régió-szálról)
- EconomyEventManager: market-boom a sokk kikapcsolásakor is fut
- SpyManager.shutdown: élő álcák levétele disable/reloadkor
- ServerChallengeManager.updateBar synchronized (bossbar-verseny)
- Baráti tűz: AngryChicken + Shadowburn isAlly-szűrés
- UndeadUtil = egyetlen élőhalott-definíció (SpellTargetingUtil delegál)

APRÓ + DEAD:
- DonationChest debounce-mentés; ambient-jutalom AFK-fék; TerritoryProtection
  null-guard; ResourceManager holt ctor-paraméter; CurrencyItemFactory holt
  overload; 6 hívó nélküli metódus törölve (RelicManager.describeDefinitions,
  StrangerNpc.isPresent, EconomyEvent.getEventCurrency, Guild.hasPendingInvite,
  Armament.isArmamentActive, SeasonMonument.engravedLines, Hud.label)
- PLAYTEST: audit-blokk

+++ src/main/java/hu/taliann/icesmp/managers/ClaimManager.java
@@ -492,7 +492,9 @@ public final class ClaimManager implements PersistentStore, hu.taliann.icesmp.se
         final int maxX = Math.max(selection.x1, selection.x2);
         final int minZ = Math.min(selection.z1, selection.z2);
         final int maxZ = Math.max(selection.z1, selection.z2);
-        final int columns = (maxX - minX + 1) * (maxZ - minZ + 1);
+        // long-szorzás: egy ~46341×46341-es kijelölésnél az int-szorzat átfordulna, és a
+        // méret-plafon + a költség-számítás is kijátszható lenne.
+        final long columns = (long) (maxX - minX + 1) * (long) (maxZ - minZ + 1);
         final int areaMax = Math.max(16, configManager.getInt("claims.area-max-columns", 6400));
         if (columns > areaMax) {
             return "claim-area-too-big";
